refactor(exception): move exceptions-related classes into dedicated package
