+++ tracker-api/src/main/java/com/sloyardms/trackerapi/common/exception/ResourceDuplicatedException.java
@@ -1,4 +1,4 @@
-package com.sloyardms.trackerapi.exception;
+package com.sloyardms.trackerapi.common.exception;
 
 public class ResourceDuplicatedException extends RuntimeException{
