+++ tracker-api/src/main/java/com/sloyardms/trackerapi/group/GroupServiceImpl.java
@@ -5,8 +5,8 @@ import com.sloyardms.trackerapi.group.dto.GroupDto;
 import com.sloyardms.trackerapi.group.dto.GroupUpdateDto;
 import com.sloyardms.trackerapi.group.entity.Group;
 import com.sloyardms.trackerapi.user.entity.User;
-import com.sloyardms.trackerapi.exception.ResourceDuplicatedException;
-import com.sloyardms.trackerapi.exception.ResourceNotFoundException;
+import com.sloyardms.trackerapi.common.exception.ResourceDuplicatedException;
+import com.sloyardms.trackerapi.common.exception.ResourceNotFoundException;
 import com.sloyardms.trackerapi.group.mapper.GroupMapper;
 import com.sloyardms.trackerapi.user.UserRepository;
 import org.springframework.data.domain.Page;
