+++ tracker-api/src/main/java/com/sloyardms/trackerapi/tag/TagServiceImpl.java
@@ -5,8 +5,8 @@ import com.sloyardms.trackerapi.tag.dto.TagDto;
 import com.sloyardms.trackerapi.tag.dto.TagUpdateDto;
 import com.sloyardms.trackerapi.tag.entity.Tag;
 import com.sloyardms.trackerapi.user.entity.User;
-import com.sloyardms.trackerapi.exception.ResourceDuplicatedException;
-import com.sloyardms.trackerapi.exception.ResourceNotFoundException;
+import com.sloyardms.trackerapi.common.exception.ResourceDuplicatedException;
+import com.sloyardms.trackerapi.common.exception.ResourceNotFoundException;
 import com.sloyardms.trackerapi.tag.mapper.TagMapper;
 import com.sloyardms.trackerapi.user.UserRepository;
 import org.springframework.data.domain.Page;
