+++ tracker-api/src/main/java/com/sloyardms/trackerapi/user/UserServiceImpl.java
@@ -4,8 +4,8 @@ import com.sloyardms.trackerapi.user.dto.UserCreateDto;
 import com.sloyardms.trackerapi.user.dto.UserDto;
 import com.sloyardms.trackerapi.user.dto.UserUpdateDto;
 import com.sloyardms.trackerapi.user.entity.User;
-import com.sloyardms.trackerapi.exception.ResourceDuplicatedException;
-import com.sloyardms.trackerapi.exception.ResourceNotFoundException;
+import com.sloyardms.trackerapi.common.exception.ResourceDuplicatedException;
+import com.sloyardms.trackerapi.common.exception.ResourceNotFoundException;
 import com.sloyardms.trackerapi.user.mapper.UserMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
