+++ tracker-api/src/test/java/com/sloyardms/trackerapi/controller/UserControllerSliceTests.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sloyardms.trackerapi.user.UserController;
 import com.sloyardms.trackerapi.user.dto.UserCreateDto;
 import com.sloyardms.trackerapi.user.dto.UserUpdateDto;
-import com.sloyardms.trackerapi.exception.ResourceDuplicatedException;
-import com.sloyardms.trackerapi.exception.ResourceNotFoundException;
+import com.sloyardms.trackerapi.common.exception.ResourceDuplicatedException;
+import com.sloyardms.trackerapi.common.exception.ResourceNotFoundException;
 import com.sloyardms.trackerapi.user.UserService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
